fix(paint): 显式浮点转换，修复 clang CI 的 -Wdouble-promotion 错误
graphics::OutlineEdge（float）转换到 image::SvgOutlineEdge（double）处，GCC 不告警而 clang 报 implicit conversion increases floating-point precision；CI 启用 NEKO_WARNINGS_AS_ERRORS=ON 后成为编译错误。ConvertEdge() 的 p[i] 与 CreateSvgTextShaper() 的 advance 改为显式 static_cast<double>。无行为变化，仅将既有转换显式化。

验证：clang++ -fsyntax-only 加 CI 同款警告集（含 -Werror）通过；以 compile_commands.json 对全部 144 个项目编译单元做 clang 重扫，无其他告警。

+++ src/paint/src/svg_text_shaper.cpp
@@ -53,7 +53,7 @@ image::SvgOutlineEdge ConvertEdge(const graphics::OutlineEdge& edge)
     break;
   }
   for (int i = 0; i < 6; ++i) {
-    out.p[static_cast<std::size_t>(i)] = edge.p[i];
+    out.p[static_cast<std::size_t>(i)] = static_cast<double>(edge.p[i]);
   }
   return out;
 }
@@ -109,7 +109,7 @@ image::SvgTextShaper CreateSvgTextShaper(const graphics::FontRegistry& fonts)
         return false;
       }
       image::SvgGlyphOutline glyph;
-      glyph.advance = outline->advance;
+      glyph.advance = static_cast<double>(outline->advance);
       glyph.edges.reserve(outline->edges.size());
       for (const graphics::OutlineEdge& edge : outline->edges) {
         glyph.edges.push_back(ConvertEdge(edge));
